requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.32.4
Pillow==10.3.0
Werkzeug==3.0.6
pytest==7.4.3
pytest-flask==1.3.0